+++ src/main/java/org/opentripplanner/api/model/ApiRouterInfo.java
@@ -7,6 +7,7 @@ import org.opentripplanner.api.mapping.TraverseModeMapper;
 import org.opentripplanner.common.geometry.GeometrySerializer;
 import org.opentripplanner.routing.vehicle_rental.VehicleRentalStationService;
 import org.opentripplanner.routing.graph.Graph;
+import org.opentripplanner.routing.vehicle_parking.VehicleParkingService;
 import org.opentripplanner.util.TravelOption;
 import org.opentripplanner.util.TravelOptionsMaker;
 import org.opentripplanner.util.WorldEnvelope;
@@ -40,14 +41,22 @@ public class ApiRouterInfo {
 
     public boolean hasBikeSharing;
 
-    public boolean hasBikePark;
+    public final boolean hasBikePark;
+
+    public final boolean hasCarPark;
+
+    public final boolean hasVehicleParking;
 
     public List<TravelOption> travelOptions;
 
 
+    /** TODO: Do not pass in the graph here, do this in a mapper instead. */
     public ApiRouterInfo(String routerId, Graph graph) {
-        VehicleRentalStationService service = graph.getService(
+        VehicleRentalStationService vehicleRentalService = graph.getService(
                 VehicleRentalStationService.class, false
+        );
+        VehicleParkingService vehicleParkingService = graph.getService(
+            VehicleParkingService.class, false
         );
 
         this.routerId = routerId;
@@ -58,8 +67,10 @@ public class ApiRouterInfo {
         this.transitModes = TraverseModeMapper.mapToApi(graph.getTransitModes());
         this.envelope = graph.getEnvelope();
         this.hasParkRide = graph.hasParkRide;
-        this.hasBikeSharing = mapHasBikeSharing(service);
-        this.hasBikePark = mapHasBikePark(service);
+        this.hasBikeSharing = mapHasBikeSharing(vehicleRentalService);
+        this.hasBikePark = mapHasBikePark(vehicleParkingService);
+        this.hasCarPark = mapHasCarPark(vehicleParkingService);
+        this.hasVehicleParking = mapHasVehicleParking(vehicleParkingService);
         this.travelOptions = TravelOptionsMaker.makeOptions(graph);
         addCenter(graph.getCenter());
     }
@@ -73,11 +84,25 @@ public class ApiRouterInfo {
         return service.getVehicleRentalPlaces().size() > 1;
     }
 
-    public boolean mapHasBikePark(VehicleRentalStationService service) {
+    public boolean mapHasBikePark(VehicleParkingService service) {
         if (service == null) {
             return false;
         }
-        return !service.getBikeParks().isEmpty();
+        return service.getBikeParks().findAny().isPresent();
+    }
+
+    public boolean mapHasCarPark(VehicleParkingService service) {
+        if (service == null) {
+            return false;
+        }
+        return service.getCarParks().findAny().isPresent();
+    }
+
+    public boolean mapHasVehicleParking(VehicleParkingService service) {
+        if (service == null) {
+            return false;
+        }
+        return service.getVehicleParkings().findAny().isPresent();
     }
 
     /**