+++ src/main/java/org/opentripplanner/routing/api/request/RoutingRequest.java
@@ -236,7 +236,6 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
     public double maxWheelchairSlope = 0.0833333333333; // ADA max wheelchair ramp slope is a good default.
 
     /** Whether the planner should return intermediate stops lists for transit legs. */
-    // TODO OTP2 Maybe this should be up to the API?
     public boolean showIntermediateStops = false;
 
     /** max walk/bike speed along streets, in meters per second */
@@ -354,6 +353,12 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
 
     /** Cost of parking a bike. */
     public int bikeParkCost = 120;
+
+    /** Time to park a car */
+    public int carParkTime = 60;
+
+    /** Cost of parking a car. */
+    public int carParkCost = 120;
 
     /**
      * Time to park a car in a park and ride, w/o taking into account driving and walking cost
@@ -648,7 +653,6 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
       This is a temporary solution, as it only covers parking and rental at the beginning of the trip.
     */
     public boolean bikeRental = false;
-    public boolean bikeParkAndRide = false;
     public boolean parkAndRide  = false;
     public boolean carPickup = false;
 
@@ -1008,7 +1012,7 @@ public class RoutingRequest implements AutoCloseable, Cloneable, Serializable {
                     break;
                 case BIKE_TO_PARK:
                     streetRequest.setStreetSubRequestModes(new TraverseModeSet(TraverseMode.BICYCLE, TraverseMode.WALK));
-                    streetRequest.bikeParkAndRide = true;
+                    streetRequest.parkAndRide = true;
                     break;
                 case BIKE_RENTAL:
                     streetRequest.setStreetSubRequestModes(new TraverseModeSet(TraverseMode.BICYCLE, TraverseMode.WALK));