+++ src/test/java/org/opentripplanner/graph_builder/module/VehicleParkingLinkingTest.java
@@ -1,0 +1,177 @@
+package org.opentripplanner.graph_builder.module;
+
+import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertTrue;
+
+import java.util.HashMap;
+
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import org.opentripplanner.common.geometry.GeometryUtils;
+import org.opentripplanner.graph_builder.DataImportIssueStore;
+import org.opentripplanner.model.FeedScopedId;
+import org.opentripplanner.routing.edgetype.StreetEdge;
+import org.opentripplanner.routing.edgetype.StreetTraversalPermission;
+import org.opentripplanner.routing.edgetype.StreetVehicleParkingLink;
+import org.opentripplanner.routing.edgetype.VehicleParkingEdge;
+import org.opentripplanner.routing.graph.Graph;
+import org.opentripplanner.routing.vehicle_parking.VehicleParking;
+import org.opentripplanner.routing.vehicle_parking.VehicleParkingHelper;
+import org.opentripplanner.routing.vehicle_parking.VehicleParkingService;
+import org.opentripplanner.routing.vertextype.IntersectionVertex;
+import org.opentripplanner.routing.vertextype.StreetVertex;
+import org.opentripplanner.routing.vertextype.VehicleParkingEntranceVertex;
+
+public class VehicleParkingLinkingTest {
+
+  private Graph graph;
+  private IntersectionVertex A, B;
+
+  @BeforeEach
+  public void setup() {
+    graph = new Graph();
+    graph.hasStreets = true;
+
+    A = new IntersectionVertex(graph, "A", 0, 0);
+    B = new IntersectionVertex(graph, "B", 0.01, 0);
+
+    street(A, B, StreetTraversalPermission.PEDESTRIAN);
+  }
+
+  @Test
+  public void entranceWithVertexLinkingTest() {
+    var parking = VehicleParking.builder()
+            .entrance(builder -> builder
+                    .entranceId(new FeedScopedId("TEST", "1"))
+                    .vertex(A))
+            .build();
+    var parkingVertex = new VehicleParkingEntranceVertex(graph, parking.getEntrances().get(0));
+
+    StreetLinkerModule streetLinkerModule = new StreetLinkerModule();
+    streetLinkerModule.buildGraph(graph, new HashMap<>(), new DataImportIssueStore(false));
+
+    assertEquals(1, parkingVertex.getOutgoing().size());
+    parkingVertex.getOutgoing().forEach(e -> assertEquals(e.getToVertex(), A));
+
+    assertEquals(1, parkingVertex.getIncoming().size());
+    parkingVertex.getIncoming().forEach(e -> assertEquals(e.getFromVertex(), A));
+  }
+
+  @Test
+  public void entranceWithoutVertexLinkingTest() {
+    var parking = VehicleParking.builder()
+            .entrance(builder -> builder
+                    .entranceId(new FeedScopedId("TEST", "1"))
+                    .x(0.0001)
+                    .y(0)
+                    .carAccessible(true)
+                    .walkAccessible(true))
+            .build();
+    var parkingVertex = new VehicleParkingEntranceVertex(graph, parking.getEntrances().get(0));
+
+    StreetLinkerModule streetLinkerModule = new StreetLinkerModule();
+    streetLinkerModule.buildGraph(graph, new HashMap<>(), new DataImportIssueStore(false));
+
+    var streetLinks = graph.getEdgesOfType(StreetVehicleParkingLink.class);
+    assertEquals(2, streetLinks.size());
+
+    streetLinks.forEach(e -> assertTrue(e.getFromVertex().equals(parkingVertex) ^ e.getToVertex().equals(parkingVertex)));
+  }
+
+  @Test
+  public void carParkingEntranceToAllTraversableStreetLinkingTest() {
+    var C = new IntersectionVertex(graph, "C", 0.0001, 0.0001);
+    var D = new IntersectionVertex(graph, "D", 0.01, 0.01);
+    street(C, D, StreetTraversalPermission.CAR);
+
+    street(A, C, StreetTraversalPermission.NONE);
+
+    var parking = VehicleParking.builder()
+            .entrance(builder -> builder
+                    .entranceId(new FeedScopedId("TEST", "1"))
+                    .x(0.0001)
+                    .y(0)
+                    .carAccessible(true)
+                    .walkAccessible(true))
+            .build();
+    var parkingVertex = new VehicleParkingEntranceVertex(graph, parking.getEntrances().get(0));
+
+    StreetLinkerModule streetLinkerModule = new StreetLinkerModule();
+    streetLinkerModule.buildGraph(graph, new HashMap<>(), new DataImportIssueStore(false));
+
+    var streetLinks = graph.getEdgesOfType(StreetVehicleParkingLink.class);
+    assertEquals(4, streetLinks.size());
+
+    streetLinks.forEach(e -> assertTrue(e.getFromVertex().equals(parkingVertex) ^ e.getToVertex().equals(parkingVertex)));
+  }
+
+  private static void street(StreetVertex from, StreetVertex to, StreetTraversalPermission permissions) {
+    new StreetEdge(from, to,
+        GeometryUtils.makeLineString(from.getLat(), from.getLon(), to.getLat(), to.getLon()),
+        String.format("%s%s street", from.getName(), to.getName()),
+        1,
+        permissions,
+        false
+    );
+  }
+
+  @Test
+  public void removeEntranceWithNonExistingVertexTest() {
+    var vehicleParking = VehicleParking.builder()
+        .id(new FeedScopedId("TEST", "VP"))
+        .bicyclePlaces(true)
+        .entrance(builder -> builder
+                .entranceId(new FeedScopedId("TEST", "Entrance-1"))
+                .vertex(A)
+                .walkAccessible(true))
+        .entrance(builder -> builder
+                .entranceId(new FeedScopedId("TEST", "Entrance-2"))
+                .vertex(B)
+                .walkAccessible(true))
+        .build();
+
+    VehicleParkingHelper.linkVehicleParkingToGraph(graph, vehicleParking);
+
+    graph.remove(A);
+
+    StreetLinkerModule streetLinkerModule = new StreetLinkerModule();
+    streetLinkerModule.buildGraph(graph, new HashMap<>(), new DataImportIssueStore(false));
+
+    assertEquals(1, vehicleParking.getEntrances().size());
+
+    assertEquals(1, graph.getVerticesOfType(VehicleParkingEntranceVertex.class).size());
+
+    assertEquals(1, graph.getEdgesOfType(VehicleParkingEdge.class).size());
+    assertEquals(2, graph.getEdgesOfType(StreetVehicleParkingLink.class).size());
+  }
+
+  @Test
+  public void removeVehicleParkingWithOneEntranceAndNonExistingVertexTest() {
+
+    var vehicleParking = VehicleParking.builder()
+        .id(new FeedScopedId("TEST", "VP"))
+        .bicyclePlaces(true)
+        .entrance(builder -> builder
+                .entranceId(new FeedScopedId("TEST", "Entrance-1"))
+                .vertex(A)
+                .walkAccessible(true))
+        .build();
+
+    var vehicleParkingService = graph.getService(VehicleParkingService.class, true);
+    vehicleParkingService.addVehicleParking(vehicleParking);
+    VehicleParkingHelper.linkVehicleParkingToGraph(graph, vehicleParking);
+
+    graph.remove(A);
+
+    StreetLinkerModule streetLinkerModule = new StreetLinkerModule();
+    streetLinkerModule.buildGraph(graph, new HashMap<>(), new DataImportIssueStore(false));
+
+    assertEquals(0, graph.getVerticesOfType(VehicleParkingEntranceVertex.class).size());
+
+    assertEquals(0, graph.getEdgesOfType(VehicleParkingEdge.class).size());
+    assertEquals(0, graph.getEdgesOfType(StreetVehicleParkingLink.class).size());
+
+    assertEquals(0, graph.getEdgesOfType(StreetVehicleParkingLink.class).size());
+    assertEquals(0, vehicleParkingService.getVehicleParkings().count());
+  }
+}